% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioData.R
\docType{data}
\name{bioData}
\alias{bioData}
\title{Biological data...}
\format{A data frame with 30 rows and 49 variables: ...}
\source{
\url{??? Is there a source paper?}
}
\usage{
bioData
}
\description{
...
}
\keyword{datasets}

