% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdm.map.R
\name{gdm.map}
\alias{gdm.map}
\title{Function to map NMDS transformed gdm site pair dissimilarity predictions}
\usage{
gdm.map(spData, predMap, model, k = 3, t = 0.1)
}
\arguments{
\item{spData}{Compiled dataset as output from "format.site.pair".}

\item{predMap}{Raster object with the same set of predictior variables as used for gdm model building; if no raster object is provided NMDS scores will be returned as data.frame.}

\item{model}{A gdm.model object for dissimilarity prediction.}

\item{k}{number of resulting NMDS components (default = 3); if k = 0 number of components will be derived using a mean NMDS stress value threshold (t) after 20 iterations of NMDS using all given predictor variables.}

\item{t}{NMDS stress value threshold (default = 0.1).}
}
\value{
NMDS object with k number of components based on gdm predicted sample pair dissimilarities if no raster is provided; NMDS raster map with k number of layers.
}
\description{
Function to predict dissimilarities between sample site pairs using a gdm model. Resulting multidimesional dissimilarities
             are transformed to a NMDS (non-metrical multidimensional scaling) feature space with a defined number of components.
             If a raster object with the same set of predictor variables as used for gdm model building is provided, the transformed
             dissimilarities will be mapped using k-nearest neighbor imputation.
}

