% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdm.validate.R
\name{gdm.validate}
\alias{gdm.validate}
\title{Performs (independent) validation of GDM model}
\usage{
gdm.validate(caldata, valdata, performance = "r2", geo = F)
}
\arguments{
\item{caldata}{Calibration dataset, in combined site pair ("spData") data format.}

\item{valdata}{Validation dataset, in combined site pair ("spData") data format.}

\item{performance}{Performance metric to be used for validation: \code{"rmse"} for root mean square error (RMSE) or \code{"r2"} for coefficient of determination (r2). Set as \code{"r2"} per default.}

\item{geo}{Optional use of geographical distance as predictor in GDM model. set as \code{FALSE} per default.}
}
\value{
Returns model performance value.
}
\description{
This function performs (independent) validation of a GDM model.

It requires (independent) calibration and validation datasets in combined site pair ("spData") data format, as well as the definition of the performance measure to be used in the validation and the optional use of geographical distance as predictor variable in the gdm.

#' For more details relating to "spData" data format, check \code{gdm} package.
}

