% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdm.varsig.R
\name{gdm.varsig}
\alias{gdm.varsig}
\title{Calculates statistical significance of predictor variable contributions in GDM model}
\usage{
gdm.varsig(predData, bioData, metric = "bray", geo = F, perm = 100,
  sig = 0.05)
}
\arguments{
\item{predData}{Predictor dataset ("predData" format).}

\item{bioData}{Biological dataset ("bioData" format).}

\item{metric}{Dissimilarity metric to be used.}

\item{geo}{Optional use of geographical distance as predictor in GDM model. Set to \code{FALSE} per default}

\item{perm}{Number of matrix permutations to be used for calculating statistical significance.}

\item{sig}{Significance level (p-value) to be tested. Set as 0.05 per default.}
}
\value{
Returns vector resulting of the significance test. Variables with significant model contributions are assigned value \code{TRUE}, and non-significant variables the value \code{FALSE}, according to the defined p-value threshold.
}
\description{
This function calculates the statistical significance of predictor variable contributions in GDM model, as calculated with the \code{gdm.varcont} function.

Significance is calculated through random matrix permutations, to infer if the variable (drop) contributions are greater than those obtained by chance.

It requires a biological ("bioData" format) and a predictor ("spData" format) datasets, the dissimilarity metric to be used, the optional use of geographical distance as predictor variable in the GDM, the number of permutations to be run and the significance level to be tested.

This current implementation only allows biological data in the format 1 using abundance values, as described in the \code{gdm} package.

For more details relating to "bioData" and "predData" data formats, check \code{gdm} package.
}

