% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predData.transform.R
\name{predData.transform}
\alias{predData.transform}
\title{Function to perform SCCA transformation of environmental data}
\usage{
predData.transform(predData, v.best)
}
\arguments{
\item{predData}{Environmental data as data frame (Format: Plot_ID, X, Y, env.data) or as a raster object}

\item{v.best}{sparse canonical vectors as out from \code {sgdm.best}}
}
\value{
Returns CCA transformed environmental data
}
\description{
This function transforms the given environmental input data into sparse canonical components,
             based on the sparse canonical vectors as derived from \code {sgdm.best}.Input data can either be provided
             as data frame or raster object.Input data must have the same number of predictors as used for deriving the sparse canonical vectors
}

