% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra.R
\docType{data}
\name{spectra}
\alias{spectra}
\title{Predictor data...}
\format{A data frame with 30 rows and 86 variables: ...}
\source{
\url{??? Is there a source paper?}
}
\usage{
spectra
}
\description{
...
}
\keyword{datasets}

