% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.image.R
\docType{data}
\name{spectral.image}
\alias{spectral.image}
\title{Spectral image}
\format{A raster stack with 83 layers.}
\usage{
spectral.image
}
\description{
Example predictor map for use in SGDM. Consists of a hyperspectral imagery with 83 spectral narrow-bands, and pixel size of 30m.
}
\keyword{datasets}

