% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predMap.R
\docType{data}
\name{predMap}
\alias{predMap}
\title{Predictor map...}
\format{A raster stack with 83 layers:}
\source{
\url{??? Is there a source paper?}
}
\usage{
predMap
}
\description{
...
}
\keyword{datasets}

