% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\docType{data}
\name{trees}
\alias{trees}
\title{Cerrado trees}
\format{A data frame with 30 rows and 49 columns.}
\usage{
trees
}
\description{
Example biological dataset for use in SGDM. Consists of 30 observations (sites/pixels) with abundance values for 48 tree families in the Brazilian Cerrado.
}
\keyword{datasets}

