% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.reduce.R
\name{data.reduce}
\alias{data.reduce}
\title{Reduces predictor data based on significance test}
\usage{
data.reduce(data, datatype = "sp", sigtest)
}
\arguments{
\item{data}{Data to be reduced based on predictor variable significance. It can be either a predictor dataset ("predData" format) or combined site pair table ("spData" format).}

\item{datatype}{Type of data to de reduced: \code{pred} for predictor data ("predData" format) or \code{sp} for site pair data ("spData" format).}

\item{sigtest}{Predictor variable contribution significance test, as output by gdm.varsig function.}
}
\value{
Returns reduced environmental data in the same format as the input data.
}
\description{
This function reduces the predictor variables of a dataset based on the significance test, resulting from function gdm.varsig. It can be applied to either predictor or site pair datasets.

abcd

abc
}

