% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdm.cv.R
\name{gdm.cv}
\alias{gdm.cv}
\title{Function to perform n-fold cross-validation of gdm model}
\usage{
gdm.cv(spData, nfolds = 10, metric = "bray", performance = "rmse",
  uncertainty = F, geo = F)
}
\arguments{
\item{spData}{Compiled dataset as output from "format.site.pair" function}

\item{nfolds}{Number of folds for cross-validation; default = 10; if nfolds = samples then performs leave-one-out cross-validation}
}
\value{
ABC
}
\description{
Function to perform n-fold cross-validation of gdm model
}
\examples{
gdm.cv(...)
}

