% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdm.validate.R
\name{gdm.validate}
\alias{gdm.validate}
\title{Function to calculate predictor variable drop contributions on GDM model}
\usage{
gdm.validate(caldata, valdata, performance = "r2", geo = F)
}
\arguments{
\item{caldata}{Calibration data}

\item{valdata}{Validation data}

\item{performance}{Performance metric to be used ("rmse" or "r2"), set as "r2" per default}

\item{geo}{Optional use of geographical distance as predictor in GDM model, set as FALSE per default}
}
\value{
Returns model performance value
}
\description{
Function to calculate predictor variable drop contributions on GDM model
}
\examples{
gdm.validate(...)
}

