% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdm.varcont.R
\name{gdm.varcont}
\alias{gdm.varcont}
\title{Function to calculate predictor variable drop contributions on GDM model}
\usage{
gdm.varcont(spData, geo = FALSE)
}
\arguments{
\item{spData}{Site pair table as from formatsitetable function in gdm package}

\item{geo}{Optional use of geographical distance as predictor in GDM model, set as FALSE per default}
}
\value{
Returns variable contribution in percentage
}
\description{
...
}

