% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predData.transform.R
\name{predData.transform}
\alias{predData.transform}
\title{Performs SCCA transformation of environmental data}
\usage{
predData.transform(predData, v)
}
\arguments{
\item{predData}{Predictor dataset (predData format) or as a raster object.}

\item{v}{Sparse canonical vectors as extracted by \code{sgdm.best} using output = \code{"v"}.}
}
\value{
Returns environmental data transformed into sparse canonical components for further use with GDM. This dataset is delivered in the same format (data frame or raster object) as the input data.
}
\description{
This function transforms a predictor dataset into sparse canonical components, based on the sparse canonical vectors extracted by function \code{sgdm.best} using output = \code{"v"}.

The input data can either be provided as dataframe or raster object, which must have the same number of predictors as used for deriving the sparse canonical vectors. The output data will be delivered in the same format (dataframe or raster object) as the input.

If predictor dataset is a raster object, this function requires the previous instalation of the \code{raster} package.

For more details relating to "predData" and "spData" data formats, check \code{gdm} package.
}

