% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predData.transform.R
\name{predData.transform}
\alias{predData.transform}
\title{Function for ...}
\usage{
predData.transform(predData, bioData, perf.matrix, comps = 10)
}
\arguments{
\item{predData}{...}

\item{bioData}{...}

\item{perf.matrix}{...}

\item{comps}{...}
}
\value{
Returns ...
}
\description{
...
}

