% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgdm.best.R
\name{sgdm.best}
\alias{sgdm.best}
\title{Function to retrieve the best SGDM model, SCCA canonical components or SCCA canonical vectors, as resulting from the SCCA parameter estimation using sgdm.gridsearch}
\usage{
sgdm.best(perf.matrix, predData, bioData, output = "m", comps = 10,
  metric = "bray", geo = F)
}
\arguments{
\item{perf.matrix}{Performance matrix as output from sgdm.grid}

\item{predData}{Environmental data matrix, with Plot_ID, X, Y as three first columns followed by predictor values per plot}

\item{bioData}{Biological data matrix, with Plot_ID as first column, followed by species occurrence / abundance per plot}

\item{output}{Type of output: "m" = gdm model; "c" = sparse canonical components; "v" = sparse canonical vectors; default = gdm model}

\item{comps}{Number of sparce canonical components to be calculated, set as 10 per default}

\item{metric}{Only needed if output = "m"; dissimilarity metric to be used ("bray curtis" for abundance or "Jaccard" for presence-absence), set as "bray curtis" per default}

\item{geo}{only needed if output = "m"; optional use of geographical distance as predictor in GDM model, set as FALSE per default}
}
\value{
Returns GDM model, sparse canonical components, or sparse canonical vectors (default is \code{output = "m"}, which returns a GDM model object)
}
\description{
Function to retrieve the best SGDM model, SCCA canonical components or SCCA canonical vectors, as resulting from the SCCA parameter estimation using sgdm.gridsearch
}

