% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgdm.train.R
\name{sgdm.train}
\alias{sgdm.train}
\title{Function to perform parameter estimation of SCCA via grid search, based on GDM leave one out cross-validated performances (RMSE)}
\usage{
sgdm.train(predData, bioData, comps = 10, metric = "bray",
  predPenalization = seq(0.6, 1, 0.1), bioPenalization = seq(0.6, 1, 0.1),
  geo = F)
}
\arguments{
\item{predData}{Environmental data matrix, with Plot_ID, X, Y as three first columns followed by predictor values per plot}

\item{bioData}{Biological data matrix, with Plot_ID as first column, followed by species occurrence / abundance per plot}

\item{comps}{Number of sparce canonical components to be calculated, set as 10 per default}

\item{metric}{Dissimilarity metric to be used ("bray curtis" for abundance or "Jaccard" for presence-absence), set as bray curtis" per default}

\item{predPenalization}{Vector with possible penalisation values to be applied on the environmental data matrix (between 0 and 1)}

\item{bioPenalization}{Vector with possible penalisation values to be applied on the biological data matrix (between 0 and 1)}

\item{geo}{Optional use of geographical distance as predictor in GDM model, set as FALSE per default}
}
\value{
Returns performance matrix with RMSE values for each SCCA penalization parameter pair
}
\description{
Function to perform parameter estimation of SCCA via grid search, based on GDM leave one out cross-validated performances (RMSE)
}

