% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predData.R
\docType{data}
\name{predData}
\alias{predData}
\title{Predictor data...}
\format{A data frame with 30 rows and 86 variables: ...}
\source{
\url{??? Is there a source paper?}
}
\usage{
predData
}
\description{
...
}
\keyword{datasets}

