% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra.R
\docType{data}
\name{spectra}
\alias{spectra}
\title{Spectral dataset}
\format{A data frame with 30 rows and 86 variables.}
\usage{
spectra
}
\description{
Example predictor dataset for use in SGDM. Consists of a hyperspectral dataset with 83 spectral narrow-bands, for the 30 sampled sites (see \code{Cerrado trees} dataset).
}
\keyword{datasets}

