% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.image.R
\docType{data}
\name{spectral.image}
\alias{spectral.image}
\title{Predictor map...}
\format{A raster stack with 83 layers:}
\source{
\url{??? Is there a source paper?}
}
\usage{
spectral.image
}
\description{
...
}
\keyword{datasets}

